backend/requirements.txt
flask==3.0.0
flask-sqlalchemy==3.1.1
flask-migrate==4.0.5
flask-jwt-extended==4.6.0
mysql-connector-python==8.3.0
python-dotenv==1.0.0
reportlab==4.1.0
pytest==7.4.3
flask-cors==4.0.0